Moved class Ensure to package roperty, since package commons.util only contains one class with one method now.
